- SoundUIElement refactoring - Added check empty sound clip
