+++ Assets/Scripts/Audio/UI/Elements/SoundUIElement.cs
@@ -1,33 +1,6 @@
-﻿using UnityEngine;
-using Zenject;
-
-namespace Audio
+﻿namespace Audio
 {
-    public class SoundUIElement : MonoBehaviour
+    public class SoundUIElement : SoundElement<IUISounds>
     {
-        private IAudioPlayer _audioPlayer;
-        protected IUISounds Sounds { get; private set; }
-
-        protected bool IsActive => _audioPlayer != null; 
-        
-
-        [Inject]
-        public virtual void Construct(IAudioPlayer audioPlayer, IUISounds uiSoundSettings)
-        {
-            _audioPlayer = audioPlayer;
-            Sounds = uiSoundSettings;
-        }
-
-        protected void Play(IAudioClip oneShot)
-        {
-            if (!IsActive)
-            {
-                Debug.LogError($"AudioPlayer is null. Can't play sound: {oneShot}");
-                return;
-            }
-            
-
-            _audioPlayer.PlayOneShoot(oneShot);
-        }
     }
 }
